requirements.txt

# basics
attrs==25.4.0
cached-property==1.5.2
permacache==4.1.0
Pillow==9.4.0
matplotlib==3.5.1
pdf2image==1.16.0
editdistance==0.6.2

# ml packages
numpy==1.23.3
torch==2.7.1
scipy==1.8.1
librosa==0.9.2
resampy==0.3.1

# only being used for the MNIST dataset
keras==2.12.0
tensorflow==2.12.0

pandas==1.3.5

# data & utilities
dconstruct==1.0.0
sparling
more-itertools==9.1.0
h5py==3.8.0
tqdm==4.65.0

# testing
pytest==7.4.0

# debugging
IPython
